+++ NSeleneTests/Integration/SharedDriver/SeleneElement_Find_Should.cs
@@ -4,6 +4,7 @@ using OpenQA.Selenium;
 
 namespace NSelene.Tests.Integration.SharedDriver.SeleneElementSpec
 {
+    using System;
     using Harness;
 
     [TestFixture]
@@ -144,7 +145,7 @@ namespace NSelene.Tests.Integration.SharedDriver.SeleneElementSpec
             try {
                 S("p").Find("a").Click();
                 Assert.Fail("should fail on timeout before can be clicked");
-            } catch (WebDriverTimeoutException) {
+            } catch (TimeoutException) {
                 Assert.IsFalse(Configuration.Driver.Url.Contains("second"));
             }
         }