+++ CHANGELOG.md
@@ -56,8 +56,12 @@
 - tuned selene elements representation in error messages
   - now code like `SS(".parent").FilterBy(Be.Visible)[0].SS(".child").FindBy(Have.CssClass("special")).S("./following-sibling::*")`
   - renders to: `Browser.All(.parent).By(Visible)[0].All(.child).FirstBy(has CSS class 'special').Element(./following-sibling::*)`
-- Deprecated (Marked as Obsolete)
+- improved waiting (waits not just for visibility but till "being passed") at SeleneElement's:
+  - Click()
+- *Deprecated (Marked as Obsolete)*
   - Configuration.WebDriver (use Configuration.Driver instead)
+- **potential breaking changes**:
+  - Switched to System.TimeoutException in some waits (instead of WebDriverTimeoutException)
 
 ## 1.0.0-alpha03 (to be released on 2020.06.03)
 - added `SeleneElement#Type(string keys)`, i.e. `S(selector).Type(text)`